Move Tests for String to separate package Rename classes Update copyright
