+++ restygwt/src/test/java/org/fusesource/restygwt/StringEncoderDecoderAutodetectPlainTextTestGwt.gwt.xml
@@ -1,6 +1,6 @@
 <!--
 
-    Copyright (C) 2009-2015 the original author or authors.
+    Copyright (C) 2009-2016 the original author or authors.
     See the notice.md file distributed with this work for additional
     information regarding copyright ownership.
